chore<BE>: 사소한 수정

+++ BACKEND/src/main/java/com/ssafy/challympic/api/UserApiController.java
@@ -68,7 +68,7 @@ public class UserApiController {
         if(media == null)
             return new Result(false, HttpStatus.OK.value());
 
-        Long file_no = mediaService.saveMedia(media);
+        int file_no = mediaService.saveMedia(media);
 
 //        Media findMedia = mediaService.
