+++ BACKEND/src/main/java/com/ssafy/challympic/domain/PostLike.java
@@ -11,13 +11,13 @@ import javax.persistence.*;
 public class PostLike {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long like_no;
+    private int like_no;
 
     private Integer post_no;
 
     private Integer user_no;
 
-    public PostLike(Long like_no, Integer post_no, Integer user_no){
+    public PostLike(int like_no, Integer post_no, Integer user_no){
         this.like_no = like_no;
         this.post_no = post_no;
         this.user_no = user_no;
